+++ QStyler/styler.py
@@ -23,9 +23,10 @@ import re
 from PySide6.QtCore import Qt, Signal
 from PySide6.QtGui import QValidator
 from PySide6.QtWidgets import (QApplication, QComboBox, QGroupBox, QHBoxLayout,
-                               QLabel, QPushButton, QTableWidget,
-                               QTableWidgetItem, QVBoxLayout, QWidget)
+                               QLabel, QTableWidget, QTableWidgetItem,
+                               QVBoxLayout, QWidget)
 
+from QStyler.toolbar import ToolBar
 from QStyler.utils import blockSignals
 
 
@@ -118,7 +119,7 @@ class Table(QTableWidget):
             title = self.widget.getWidgetState()
             if not prop or prop == "":
                 return  # pragma: nocover
-            self.manager.add_sheet(title, prop, value)
+            self.manager.append_sheet(title, prop, value)
         self.setNewRow.emit()
 
     def currentSheet(self):
@@ -300,6 +301,7 @@ class StylerTab(QWidget):
     def __init__(self, parent=None):
         """Initialize the styler tab."""
         super().__init__(parent=parent)
+        self.window = parent
         self.manager = parent.manager
         self.data = self.manager.data
         self.layout = QVBoxLayout()
@@ -307,17 +309,19 @@ class StylerTab(QWidget):
         self.widget_label = QLabel("Widget(s)")
         self.control_label = QLabel("Control")
         self.state_label = QLabel("State")
-        self.plusbtn = QPushButton("+", parent=self)
-        self.minusbtn = QPushButton("-", parent=self)
         self.combo = WidgetCombo(self.data, parent=self)
         self.control_combo = ControlCombo(self.data, parent=self)
         self.state_combo = StateCombo(self.data, parent=self)
-        self.button = QPushButton("Save Theme", parent=self)
         self.table = Table(self.manager, parent=self)
+        self.toolbar = ToolBar(parent=self)
+        self.hlayout2 = QHBoxLayout()
+        self.hlayout2.addStretch(0)
+        self.hlayout2.addWidget(self.toolbar)
+        self.hlayout2.addStretch(0)
+        self.layout.addLayout(self.hlayout2)
         self.vlayout1 = QVBoxLayout()
         self.vlayout2 = QVBoxLayout()
         self.vlayout3 = QVBoxLayout()
-        self.hlayout4 = QHBoxLayout()
         self.vlayout1.addWidget(self.widget_label)
         self.vlayout1.addWidget(self.combo)
         self.vlayout2.addWidget(self.control_label)
@@ -327,17 +331,12 @@ class StylerTab(QWidget):
         self.mainGroup = QGroupBox(parent=self)
         self.hlayout = QHBoxLayout()
         self.mainGroup.setLayout(self.hlayout)
-        self.hlayout4.addWidget(self.plusbtn)
-        self.hlayout4.addWidget(self.minusbtn)
         self.hlayout.addLayout(self.vlayout1)
         self.hlayout.addLayout(self.vlayout2)
         self.hlayout.addLayout(self.vlayout3)
         self.layout.addWidget(self.mainGroup)
-        self.layout.addLayout(self.hlayout4)
         self.layout.addWidget(self.table)
-        self.layout.addWidget(self.button)
-        self.plusbtn.clicked.connect(self.add_widget_combo)
-        self.minusbtn.clicked.connect(self.minus_widget_combo)
+
         self.table.setNewRow.connect(self.addTableRow)
         self.combo.widgetChanged.connect(self.statusChanged.emit)
         self.state_combo.currentIndexChanged.connect(self.statusChanged.emit)
@@ -358,8 +357,9 @@ class StylerTab(QWidget):
         if not text:
             return  # pragma: nocover
         if len(text.split(",")) >= boxlen + 1:
+
             groupbox = GroupBox(parent=self)
-            self.layout.insertWidget(boxlen + 1, groupbox)
+            self.layout.insertWidget(boxlen + 2, groupbox)
             self.boxgroups.append(groupbox)
 
     def minus_widget_combo(self):