+++ QStyler/window.py
@@ -21,12 +21,11 @@
 import sys
 from typing import Optional
 
-from PySide6.QtWidgets import (QApplication, QMainWindow, QTabWidget,
-                               QVBoxLayout, QWidget)
+from PySide6.QtWidgets import QMainWindow, QTabWidget, QVBoxLayout, QWidget
 
 from QStyler.menubar import MenuBar
 from QStyler.styler import StylerTab
-from QStyler.utils import StyleManager, get_icon
+from QStyler.utils import Application, get_icon, get_manager
 from QStyler.widgets import WidgetsTab
 
 
@@ -54,20 +53,20 @@ class MainWindow(QMainWindow):
         self.resize(900, 700)
         self.tabWidget = QTabWidget()
         self.setWindowIcon(get_icon("QStylerIcon.png"))
-        StyleManager.window = self
-        MainWindow.manager = StyleManager()
+        self.manager = get_manager()
         self.widgets = WidgetsTab(parent=self)
         self.styler = StylerTab(parent=self)
         self.tabWidget.addTab(self.widgets, "Widgets")
         self.tabWidget.addTab(self.styler, "Style")
         self.layout.addWidget(self.tabWidget)
         self.menubar = MenuBar(self)
+        self.styler.toolbar.activate_load_item()
         self.setMenuBar(self.menubar)
 
 
 def execute():  # pragma: nocover
     """Entry point for cli and execution."""
-    app = QApplication(sys.argv)
+    app = Application(sys.argv)
     window = MainWindow()
     window.show()
     sys.exit(app.exec())